pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pubmed-downloader"
version = "0.1.0"
description = "PubMed Central article fetching and markdown conversion"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "loguru>=0.7.0",
    "tqdm>=4.67.0",
    "biopython>=1.85",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pubmed_downloader"]
